+++ windows/flutter/generated_plugins.cmake
@@ -5,7 +5,6 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   audiotags
   dynamic_color
-  flutter_secure_storage_windows
   isar_flutter_libs
   permission_handler_windows
   share_plus